fix stuff partially
