+++ src/main/java/com/stoneskies/feudalism/Commands/Feudalism.java
@@ -1,6 +1,7 @@
 package com.stoneskies.feudalism.Commands;
 
 import com.stoneskies.feudalism.Commands.Debug.DebugCommands;
+import com.stoneskies.feudalism.Commands.Ruin.RuinCommands;
 import com.stoneskies.feudalism.Util.ChatInfo;
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
@@ -14,7 +15,7 @@ public class Feudalism implements CommandExecutor {
         if(label.equalsIgnoreCase("fd")) {
             if(args.length != 0) {
                 switch(args[0]) {
-                    case "townruin":
+                    case "ruin":
                         // set the executor to be the ruin commands class
                         RuinCommands.exec(sender, args);
                         break;
