+++ src/main/java/com/stoneskies/feudalism/Util/ChatInfo.java
@@ -4,6 +4,6 @@ import org.bukkit.ChatColor;
 
 public class ChatInfo {
     public static String msg(String text) {
-        return ChatColor.translateAlternateColorCodes('&', "&6[Feudalism]&r&o " + text);
+        return ChatColor.translateAlternateColorCodes('&', "&6[Feudalism]&r " + text);
     }
 }
